+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.x branch
 ## 2.1 branch
 ### 2.1.3
+* added `createBackup()` and `createSomeBackups()` to the `TestCase` class;
 * `BackupManager::_send()` has become `getEmailInstance()`.
 
 ### 2.1.2